Tidyed up code
