+++ src/functions/sendCandidateResults/domain/competency-outcome.ts
@@ -1,0 +1,5 @@
+export enum CompetencyOutcome {
+    DF = 'DF',
+    S = 'S',
+    D = 'D',
+  }
