+++ src/functions/sendCandidateResults/domain/fault.ts
@@ -1,0 +1,4 @@
+export interface Fault {
+  name: string;
+  count: number;
+}
