+++ src/functions/sendCandidateResults/framework/di/types.ts
@@ -4,4 +4,5 @@ export const TYPES = {
   IConfigAdapter: Symbol.for('IConfigAdapter'),
   IRequestScheduler: Symbol.for('IRequestScheduler'),
   ITemplateIdProvider: Symbol.for('ITemplateIdProvider'),
+  IFaultProvider: Symbol.for('IFaultProvider'),
 };
